update chuyen tien sang dinh dang tien (1,000,000 đ)

+++ app/src/main/java/com/example/doantotnghiep/model/ArrayThongKe.java
@@ -2,9 +2,9 @@ package com.example.doantotnghiep.model;
 
 public class ArrayThongKe {
     public String TenThongKe;
-    public int TienThongKe;
+    public Double TienThongKe;
 
-    public ArrayThongKe(String TenThongKe, int TienThongKe) {
+    public ArrayThongKe(String TenThongKe, Double TienThongKe) {
         this.TenThongKe = TenThongKe;
         this.TienThongKe = TienThongKe;
     }
